Plugin: Remove support for exports
