+++ include/reframework/API.h
@@ -1,17 +1,6 @@
 #ifndef REFRAMEWORK_API_H
 #define REFRAMEWORK_API_H
 
-#ifdef REFRAMEWORK_EXPORTS
-#define REFRAMEWORK_API __declspec(dllexport)
-#else
-// don't use this unless you really want to.
-#ifdef REFRAMEWORK_IMPORTS
-#define REFRAMEWORK_API __declspec(dllimport)
-#else
-#define REFRAMEWORK_API
-#endif
-#endif
-
 #define REFRAMEWORK_PLUGIN_VERSION_MAJOR 1
 #define REFRAMEWORK_PLUGIN_VERSION_MINOR 0
 #define REFRAMEWORK_PLUGIN_VERSION_PATCH 0
@@ -48,18 +37,6 @@ typedef void (*REFLuaLockUnlockFn)();
 typedef bool (*REFOnDeviceResetFn)(REFOnDeviceResetCb);
 typedef bool (*REFOnMessageFn)(REFOnMessageCb);
 
-// Optional imports. Passed through REFrameworkPluginFunctions.
-REFRAMEWORK_API bool reframework_on_initialized(REFInitializedCb cb);
-REFRAMEWORK_API bool reframework_on_lua_state_created(REFLuaStateCreatedCb cb);
-REFRAMEWORK_API bool reframework_on_lua_state_destroyed(REFLuaStateDestroyedCb cb);
-REFRAMEWORK_API bool reframework_on_frame(REFOnFrameCb cb);
-REFRAMEWORK_API bool reframework_on_pre_application_entry(const char* name, REFOnPreApplicationEntryCb cb);
-REFRAMEWORK_API bool reframework_on_post_application_entry(const char* name, REFOnPostApplicationEntryCb cb);
-REFRAMEWORK_API void reframework_lock_lua();
-REFRAMEWORK_API void reframework_unlock_lua();
-REFRAMEWORK_API bool reframework_on_device_reset(REFOnDeviceResetCb cb);
-REFRAMEWORK_API bool reframework_on_message(REFOnMessageCb cb);
-
 typedef struct {
     int major;
     int minor;
@@ -70,7 +47,6 @@ typedef struct {
 typedef void (*REFPluginRequiredVersionFn)(REFrameworkPluginVersion*);
 
 typedef struct {
-    REFOnInitializeFn on_initialized;
     REFOnLuaStateCreatedFn on_lua_state_created;
     REFOnLuaStateDestroyedFn on_lua_state_destroyed;
     REFOnFrameFn on_frame;
